requirements.txt
langchain
langchain-chroma
langchain-community
langchain-core
langchain-openai
langchain-text-splitters
langgraph
langgraph-checkpoint
langgraph-prebuilt
langgraph-sdk
langsmith
pypdf
gTTS
pygame
openai-whisper
pyaudio
webrtcvad
numpy
scipy